refactored image block
