+++ network-api/networkapi/wagtailpages/custom_blocks/base.py
@@ -1,0 +1,10 @@
+class ImageBlock(blocks.StructBlock):
+    image = ImageChooserBlock()
+    altText = blocks.CharBlock(
+        required=True,
+        help_text='Image description (for screen readers).'
+    )
+
+    class Meta:
+        icon = 'image'
+        template = 'wagtailpages/blocks/image_block.html'
