+++ network-api/networkapi/wagtailpages/custom_blocks/image_block.py
@@ -1,0 +1,13 @@
+class AnnotatedImageBlock(ImageBlock):
+    caption = blocks.CharBlock(
+        required=False
+    )
+    captionURL = blocks.CharBlock(
+        required=False,
+        help_text='Optional URL that this caption should link out to.'
+    )
+
+    class Meta:
+        icon = 'image'
+        template = 'wagtailpages/blocks/annotated_image_block.html'
+        help_text = 'Design Guideline: Please crop images to a 16:6 aspect ratio when possible.'
